+++ extension/Cargo.toml
@@ -26,7 +26,7 @@ handlebars = "5.1.0"
 log = "0.4.21"
 pgmq = "0.29"
 pgrx = "=0.13.1"
-reqwest = {version = "0.11.18", features = ["json"] }
+reqwest = {version = "0.12.16", features = ["json"] }
 serde = { version = "1.0.173", features = ["derive"] }
 serde_json = "1.0.103"
 sqlx = { version = "0.8", features = [
