+++ extension/src/executor.rs
@@ -3,15 +3,15 @@ use pgrx::prelude::*;
 use crate::guc::BATCH_SIZE;
 use crate::init::VECTORIZE_QUEUE;
 use crate::util::get_pg_conn;
+use crate::workers::get_vectorize_meta;
 use sqlx::error::Error;
 use sqlx::postgres::PgRow;
 use sqlx::{Pool, Postgres, Row};
 use tiktoken_rs::cl100k_base;
-use vectorize_core::errors::DatabaseError;
-use vectorize_core::query::{check_input, create_batches, new_rows_query_join};
-use vectorize_core::transformers::types::Inputs;
-use vectorize_core::types::{JobMessage, JobParams, TableMethod};
-use vectorize_core::worker::base::get_vectorize_meta;
+use vectorize_core::core::errors::DatabaseError;
+use vectorize_core::core::query::{check_input, create_batches, new_rows_query_join};
+use vectorize_core::core::transformers::types::Inputs;
+use vectorize_core::core::types::{JobMessage, JobParams, TableMethod};
 
 #[pg_extern]
 pub fn batch_texts(
@@ -177,7 +177,7 @@ pub async fn get_new_updates(
     }
 }
 
-fn collapse_to_csv(strings: &[String]) -> String {
+pub fn collapse_to_csv(strings: &[String]) -> String {
     strings
         .iter()
         .map(|s| {
