+++ extension/src/workers/mod.rs
@@ -1,11 +1,19 @@
 pub mod pg_bgw;
 
-use crate::core::types;
-use crate::core::worker::base::execute_job;
+use crate::executor::collapse_to_csv;
 use anyhow::Result;
 use pgmq::{Message, PGMQueueExt};
 use pgrx::*;
 use sqlx::{Pool, Postgres};
+use tiktoken_rs::cl100k_base;
+use vectorize_core::core::errors::DatabaseError;
+use vectorize_core::core::guc;
+use vectorize_core::core::transformers::http_handler;
+use vectorize_core::core::transformers::providers;
+use vectorize_core::core::transformers::types::Inputs;
+use vectorize_core::core::types;
+use vectorize_core::core::types::{JobMessage, JobParams, VectorizeMeta};
+use vectorize_core::core::worker::ops;
 
 pub async fn run_worker(
     queue: PGMQueueExt,
@@ -57,4 +65,116 @@ pub async fn run_worker(
     // return Some(), indicating that worker consumed some message
     // any possibly more messages on queue
     Ok(Some(()))
+}
+
+// get job meta
+pub async fn get_vectorize_meta(
+    job_name: &str,
+    conn: &Pool<Postgres>,
+) -> Result<VectorizeMeta, DatabaseError> {
+    let row = sqlx::query_as!(
+        VectorizeMeta,
+        "
+        SELECT
+            job_id, name, index_dist_type, transformer, params
+        FROM vectorize.job
+        WHERE name = $1
+        ",
+        job_name.to_string(),
+    )
+    .fetch_one(conn)
+    .await?;
+    Ok(row)
+}
+
+/// processes a single job from the queue
+pub async fn execute_job(dbclient: &Pool<Postgres>, msg: Message<JobMessage>) -> Result<()> {
+    let job_meta = get_vectorize_meta(&msg.message.job_name, dbclient).await?;
+    let mut job_params: JobParams = serde_json::from_value(job_meta.params.clone())?;
+    let bpe = cl100k_base().unwrap();
+
+    let guc_configs = guc::get_guc_configs(&job_meta.transformer.source, dbclient).await;
+    // if api_key found in GUC, then use that and re-assign
+    if let Some(k) = guc_configs.api_key {
+        job_params.api_key = Some(k);
+    }
+
+    let provider = providers::get_provider(
+        &job_meta.transformer.source,
+        job_params.api_key.clone(),
+        guc_configs.service_url,
+        guc_configs.virtual_key,
+    )?;
+
+    let cols = collapse_to_csv(&job_params.columns);
+
+    let job_records_query = format!(
+        "
+    SELECT
+        {primary_key}::text as record_id,
+        {cols} as input_text
+    FROM {schema}.{relation}
+    WHERE {primary_key} = ANY ($1::{pk_type}[])",
+        primary_key = job_params.primary_key,
+        cols = cols,
+        schema = job_params.schema,
+        relation = job_params.relation,
+        pk_type = job_params.pkey_type
+    );
+
+    #[derive(sqlx::FromRow)]
+    struct Res {
+        record_id: String,
+        input_text: String,
+    }
+
+    let job_records: Vec<Res> = sqlx::query_as(&job_records_query)
+        .bind(&msg.message.record_ids)
+        .fetch_all(dbclient)
+        .await?;
+
+    let inputs: Vec<Inputs> = job_records
+        .iter()
+        .map(|row| {
+            let token_estimate = bpe.encode_with_special_tokens(&row.input_text).len() as i32;
+            Inputs {
+                record_id: row.record_id.clone(),
+                inputs: row.input_text.trim().to_owned(),
+                token_estimate,
+            }
+        })
+        .collect();
+
+    let embedding_request =
+        providers::prepare_generic_embedding_request(&job_meta.transformer, &inputs);
+
+    let embeddings = provider.generate_embedding(&embedding_request).await?;
+
+    let paired_embeddings = http_handler::merge_input_output(inputs, embeddings.embeddings);
+    match job_params.clone().table_method {
+        vectorize_core::core::types::TableMethod::append => {
+            ops::update_embeddings(
+                dbclient,
+                &job_params.schema,
+                &job_params.relation,
+                &job_meta.clone().name,
+                &job_params.primary_key,
+                &job_params.pkey_type,
+                paired_embeddings,
+            )
+            .await?;
+        }
+        vectorize_core::core::types::TableMethod::join => {
+            ops::upsert_embedding_table(
+                dbclient,
+                &job_meta.name,
+                paired_embeddings,
+                "vectorize", // embeddings always live in vectorize schema in JOIN method
+                &job_params.primary_key,
+                &job_params.pkey_type,
+            )
+            .await?
+        }
+    }
+    Ok(())
 }
