+++ server/Makefile
@@ -1,7 +1,9 @@
 RUST_LOG:=info
 
-run:
-	RUST_LOG=$(RUST_LOG) cargo run
+include .env
+
+run-server:
+	RUST_LOG=$(RUST_LOG) DATABASE_URL=$(DATABASE_URL) cargo run --bin vectorize-server
 
 run-worker:
 	RUST_LOG=$(RUST_LOG) cargo run --bin vectorize-worker
