+++ server/src/core/worker/base.rs
@@ -1,17 +1,4 @@
-use crate::core::errors::DatabaseError;
-use crate::core::guc;
-use crate::core::transformers::types::Inputs;
-use crate::core::transformers::{http_handler, providers};
-use crate::core::types::{JobMessage, JobParams};
-use crate::core::worker::ops;
-
-use log::error;
-use pgmq::{Message, PGMQueueExt};
-use sqlx::{Pool, Postgres};
 use std::env;
-use tiktoken_rs::cl100k_base;
-
-use crate::core::types::VectorizeMeta;
 
 use anyhow::{Result, anyhow};
 
@@ -26,50 +13,6 @@ pub fn check_input(input: &str) -> Result<()> {
         true => Ok(()),
         false => Err(anyhow!("Invalid Input: {}", input)),
     }
-}
-
-pub fn collapse_to_csv(strings: &[String]) -> String {
-    strings
-        .iter()
-        .map(|s| {
-            check_input(s).expect("Failed to validate input");
-            s.as_str()
-        })
-        .collect::<Vec<_>>()
-        .join("|| ', ' ||")
-}
-
-pub async fn poll_job(
-    conn: &Pool<Postgres>,
-    queue: &PGMQueueExt,
-    config: &Config,
-) -> Result<Option<()>> {
-    let msg: Message<JobMessage> = match queue.read::<JobMessage>(&config.queue_name, 300_i32).await
-    {
-        Ok(Some(msg)) => msg,
-        Ok(None) => {
-            log::debug!("No message found in queue: {}", config.queue_name);
-            return Ok(None);
-        }
-        Err(e) => {
-            return Err(anyhow::anyhow!("failed reading message: {}", e));
-        }
-    };
-
-    let read_ct: i32 = msg.read_ct;
-    let msg_id: i64 = msg.msg_id;
-    if read_ct <= config.max_retries {
-        execute_job(conn, msg).await?;
-    } else {
-        error!(
-            "message exceeds max retry of {}, archiving msg_id: {}",
-            config.max_retries, msg_id
-        );
-    }
-
-    queue.archive(&config.queue_name, msg_id).await?;
-
-    Ok(Some(()))
 }
 
 #[derive(Debug, Clone)]
@@ -120,116 +63,4 @@ impl Config {
 /// source a variable from environment - use default if not exists
 pub fn from_env_default(key: &str, default: &str) -> String {
     env::var(key).unwrap_or_else(|_| default.to_owned())
-}
-
-// get job meta
-pub async fn get_vectorize_meta(
-    job_name: &str,
-    conn: &Pool<Postgres>,
-) -> Result<VectorizeMeta, DatabaseError> {
-    let row = sqlx::query_as!(
-        VectorizeMeta,
-        "
-        SELECT
-            job_id, name, index_dist_type, transformer, params
-        FROM vectorize.job
-        WHERE name = $1
-        ",
-        job_name.to_string(),
-    )
-    .fetch_one(conn)
-    .await?;
-    Ok(row)
-}
-
-/// processes a single job from the queue
-pub async fn execute_job(dbclient: &Pool<Postgres>, msg: Message<JobMessage>) -> Result<()> {
-    let job_meta = get_vectorize_meta(&msg.message.job_name, dbclient).await?;
-    let mut job_params: JobParams = serde_json::from_value(job_meta.params.clone())?;
-    let bpe = cl100k_base().unwrap();
-
-    let guc_configs = guc::get_guc_configs(&job_meta.transformer.source, dbclient).await;
-    // if api_key found in GUC, then use that and re-assign
-    if let Some(k) = guc_configs.api_key {
-        job_params.api_key = Some(k);
-    }
-
-    let provider = providers::get_provider(
-        &job_meta.transformer.source,
-        job_params.api_key.clone(),
-        guc_configs.service_url,
-        guc_configs.virtual_key,
-    )?;
-
-    let cols = collapse_to_csv(&job_params.columns);
-
-    let job_records_query = format!(
-        "
-    SELECT
-        {primary_key}::text as record_id,
-        {cols} as input_text
-    FROM {schema}.{relation}
-    WHERE {primary_key} = ANY ($1::{pk_type}[])",
-        primary_key = job_params.primary_key,
-        cols = cols,
-        schema = job_params.schema,
-        relation = job_params.relation,
-        pk_type = job_params.pkey_type
-    );
-
-    #[derive(sqlx::FromRow)]
-    struct Res {
-        record_id: String,
-        input_text: String,
-    }
-
-    let job_records: Vec<Res> = sqlx::query_as(&job_records_query)
-        .bind(&msg.message.record_ids)
-        .fetch_all(dbclient)
-        .await?;
-
-    let inputs: Vec<Inputs> = job_records
-        .iter()
-        .map(|row| {
-            let token_estimate = bpe.encode_with_special_tokens(&row.input_text).len() as i32;
-            Inputs {
-                record_id: row.record_id.clone(),
-                inputs: row.input_text.trim().to_owned(),
-                token_estimate,
-            }
-        })
-        .collect();
-
-    let embedding_request =
-        providers::prepare_generic_embedding_request(&job_meta.transformer, &inputs);
-
-    let embeddings = provider.generate_embedding(&embedding_request).await?;
-
-    let paired_embeddings = http_handler::merge_input_output(inputs, embeddings.embeddings);
-    match job_params.clone().table_method {
-        crate::core::types::TableMethod::append => {
-            ops::update_embeddings(
-                dbclient,
-                &job_params.schema,
-                &job_params.relation,
-                &job_meta.clone().name,
-                &job_params.primary_key,
-                &job_params.pkey_type,
-                paired_embeddings,
-            )
-            .await?;
-        }
-        crate::core::types::TableMethod::join => {
-            ops::upsert_embedding_table(
-                dbclient,
-                &job_meta.name,
-                paired_embeddings,
-                &job_params.schema,
-                &job_params.primary_key,
-                &job_params.pkey_type,
-            )
-            .await?
-        }
-    }
-    Ok(())
 }
